- reorganized function project - moved settings in top folder for portal - improved string substitution for keyvault to handle connection strings
