--- Portal/src/Datahub.Functions/EmailTemplates/bug_report.html
@@ -1,7 +1,0 @@
-﻿Bug Report - {title}
-<p>
-    A bug report was submitted and issue {id} was created on <a href="{url}">Azure DevOps</a>. Description: {description}
-</p>
-<p>
-    Un rapport de bogue a été soumis et le problème {id} a été créée sur <a href="{url}">Azure DevOps</a>. Description: {description}
-</p>
