--- Portal/src/Datahub.Functions/EmailTemplates/test_template.html
@@ -1,7 +1,0 @@
-﻿{key2} test
-<p>
-    {key1} test
-</p>
-<p>
-    {key1} test
-</p>
