+++ ServerlessOperations/src/Datahub.Functions/.gitignore
