+++ ServerlessOperations/src/Datahub.Functions/AzureConfig.cs
