+++ ServerlessOperations/src/Datahub.Functions/ConfigureWorkspaceAppService.cs
