+++ ServerlessOperations/src/Datahub.Functions/Datahub.Functions.csproj
