+++ ServerlessOperations/src/Datahub.Functions/EmailNotificationHandler.cs
