+++ ServerlessOperations/src/Datahub.Functions/EmailTemplates/cost_alert.html
