+++ ServerlessOperations/src/Datahub.Functions/EmailTemplates/email_template.md
