+++ ServerlessOperations/src/Datahub.Functions/EmailTemplates/project_inactive_alert.html
