+++ ServerlessOperations/src/Datahub.Functions/EmailTemplates/test_template.html
@@ -1,0 +1,7 @@
+﻿{key2} test
+<p>
+    {key1} test
+</p>
+<p>
+    {key1} test
+</p>
