+++ ServerlessOperations/src/Datahub.Functions/EmailTemplates/user_deletion_alert.html
