+++ ServerlessOperations/src/Datahub.Functions/EmailTemplates/user_invitation.html
