+++ ServerlessOperations/src/Datahub.Functions/FunctionsHealthCheck.cs
