+++ ServerlessOperations/src/Datahub.Functions/Program.cs
