+++ ServerlessOperations/src/Datahub.Functions/ProjectUsageUpdater.cs
