+++ ServerlessOperations/src/Datahub.Functions/Services/ConfigureServices.cs
