+++ ServerlessOperations/src/Datahub.Functions/TerraformOutputHandler.cs
