+++ ServerlessOperations/src/Datahub.Functions/host.json
