+++ ServerlessOperations/src/Datahub.Functions/local.settings.json-tmpl
