+++ gen-all-settings.ps1
@@ -3,5 +3,6 @@ $CurrentPath = Split-Path -Parent $MyInvocation.MyCommand.Path
 
 Import-Module $CurrentPath/scripts/appsettings.psm1 -Force
 
-Export-AppSettings -SourceFile "./ResourceProvisioner/test/ResourceProvisioner.SpecflowTests/template.appsettings.test.json" -Target AppSettings -EnvironmentName test
-Export-AppSettings -SourceFile "./Portal/src/Datahub.Portal/template.appsettings.sand.json" -Target AppSettings -EnvironmentName test
+#Export-AppSettings -SourceFile "./ResourceProvisioner/test/ResourceProvisioner.SpecflowTests/template.appsettings.test.json" -Target AppSettings -EnvironmentName test
+
+Export-AppSettings -ProjectFolder "./Portal/src/Datahub.Portal" -SourceFile "./Portal/template.appsettings.sand.json" -Target AppSettings -EnvironmentName test
